moved metadata files into metadata folder
