+++ compiler/rustc_codegen_llvm/src/gotoc/block.rs
@@ -3,7 +3,7 @@
 
 //! This file contains functions related to codegenning MIR blocks into gotoc
 
-use super::metadata::*;
+use crate::gotoc::mir_to_goto::metadata::*;
 use rustc_middle::mir::{BasicBlock, BasicBlockData};
 
 impl<'tcx> GotocCtx<'tcx> {
