+++ compiler/rustc_codegen_llvm/src/gotoc/debug.rs
@@ -3,7 +3,7 @@
 
 //! This file contains functionality that makes RMC easier to debug
 
-use super::metadata::*;
+use crate::gotoc::mir_to_goto::metadata::*;
 use rustc_middle::mir::Body;
 use rustc_middle::ty::print::with_no_trimmed_paths;
 use rustc_middle::ty::Instance;
