+++ compiler/rustc_codegen_llvm/src/gotoc/function.rs
@@ -4,7 +4,7 @@
 //! This file contains functions related to codegenning MIR functions into gotoc
 
 use super::cbmc::goto_program::{Expr, Stmt, Symbol};
-use super::metadata::*;
+use crate::gotoc::mir_to_goto::metadata::*;
 use rustc_middle::mir::{HasLocalDecls, Local};
 use rustc_middle::ty::{self, Instance, TyS};
 use tracing::{debug, warn};
