+++ compiler/rustc_codegen_llvm/src/gotoc/hooks.rs
@@ -9,9 +9,9 @@
 //! this module addresses this issue.
 
 use super::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Symbol, Type};
-use super::metadata::GotocCtx;
 use super::stubs::hash_map_stub::HashMapStub;
 use super::stubs::vec_stub::VecStub;
+use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 use rustc_hir::definitions::DefPathDataName;
 use rustc_middle::mir::{BasicBlock, Place};
 use rustc_middle::ty::print::with_no_trimmed_paths;
