+++ compiler/rustc_codegen_llvm/src/gotoc/intrinsic.rs
@@ -4,7 +4,7 @@
 use tracing::{debug, warn};
 
 use super::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Type};
-use super::metadata::GotocCtx;
+use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 use rustc_middle::mir::Place;
 use rustc_middle::ty::Instance;
 use rustc_middle::ty::{self, Ty, TyS};
