+++ compiler/rustc_codegen_llvm/src/gotoc/mir_to_goto/mod.rs
@@ -3,3 +3,5 @@
 
 pub mod monomorphize;
 //pub mod stubs;
+pub mod metadata;
+pub use metadata::GotocCtx;
