+++ compiler/rustc_codegen_llvm/src/gotoc/mod.rs
@@ -5,12 +5,10 @@ mod assumptions;
 mod backend;
 mod block;
 pub mod cbmc;
-mod current_fn;
 mod debug;
 mod function;
 mod hooks;
 mod intrinsic;
-mod metadata;
 mod mir_to_goto;
 mod operand;
 mod place;
