+++ compiler/rustc_codegen_llvm/src/gotoc/statement.rs
@@ -1,8 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 use super::cbmc::goto_program::{BuiltinFn, Expr, Location, Stmt, Type};
-use super::metadata::*;
 use super::typ::FN_RETURN_VOID_VAR_NAME;
+use crate::gotoc::mir_to_goto::metadata::*;
 use rustc_hir::def_id::DefId;
 use rustc_middle::mir;
 use rustc_middle::mir::{
