+++ compiler/rustc_codegen_llvm/src/gotoc/static_var.rs
@@ -4,7 +4,7 @@
 //! This file contains functions related to codegenning MIR static variables into gotoc
 
 use super::cbmc::goto_program::Symbol;
-use super::metadata::*;
+use crate::gotoc::mir_to_goto::metadata::*;
 use rustc_hir::def_id::DefId;
 use rustc_middle::mir::mono::MonoItem;
 use tracing::debug;
