+++ compiler/rustc_codegen_llvm/src/gotoc/stubs/hash_map_stub.rs
@@ -11,7 +11,7 @@ use std::cell::{Cell, RefCell, RefMut};
 
 use super::super::cbmc::goto_program::{Expr, Location, Stmt};
 use super::super::hooks::GotocHook;
-use super::super::metadata::GotocCtx;
+use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 
 pub struct HashMapStub<'tcx> {
     ty_opt: Cell<Option<Option<Ty<'tcx>>>>,
