+++ compiler/rustc_codegen_llvm/src/gotoc/stubs/rust_stubber.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 use super::super::cbmc::goto_program::{Expr, Location, Stmt, Type};
 use super::super::hooks::GotocTypeHook;
-use super::super::metadata::GotocCtx;
+use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 use rustc_data_structures::fx::FxHashMap;
 use rustc_hir::definitions::{DefPathData, DisambiguatedDefPathData};
 use rustc_hir::itemlikevisit::ItemLikeVisitor;
