+++ compiler/rustc_codegen_llvm/src/gotoc/typ.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 use super::cbmc::goto_program::{DatatypeComponent, Expr, Parameter, Symbol, SymbolTable, Type};
 use super::cbmc::utils::aggr_name;
-use super::metadata::GotocCtx;
 use crate::btree_map;
+use crate::gotoc::mir_to_goto::metadata::GotocCtx;
 use rustc_ast::ast::Mutability;
 use rustc_index::vec::IndexVec;
 use rustc_middle::mir::{HasLocalDecls, Local, Operand, Place, Rvalue};
